Move datamodules to datamodule sub-package
