+++ src/anomalib/data/datamodules/depth/folder_3d.py
@@ -12,7 +12,7 @@ from pandas import DataFrame, isna
 from torchvision.transforms.v2 import Transform
 
 from anomalib import TaskType
-from anomalib.data.base import AnomalibDataModule, AnomalibDepthDataset
+from anomalib.data.datamodules.base import AnomalibDataModule, AnomalibDepthDataset
 from anomalib.data.errors import MisMatchError
 from anomalib.data.utils import (
     DirType,
