+++ src/anomalib/data/datamodules/image/__init__.py
