+++ src/anomalib/data/datamodules/image/kolektor.py
@@ -27,7 +27,7 @@ from sklearn.model_selection import train_test_split
 from torchvision.transforms.v2 import Transform
 
 from anomalib import TaskType
-from anomalib.data.base import AnomalibDataModule, AnomalibDataset
+from anomalib.data.datamodules.base import AnomalibDataModule, AnomalibDataset
 from anomalib.data.errors import MisMatchError
 from anomalib.data.utils import (
     DownloadInfo,
