+++ src/anomalib/data/datamodules/image/mvtec.py
@@ -33,7 +33,7 @@ from pandas import DataFrame
 from torchvision.transforms.v2 import Transform
 
 from anomalib import TaskType
-from anomalib.data.base import AnomalibDataModule, AnomalibDataset
+from anomalib.data.datamodules.base import AnomalibDataModule, AnomalibDataset
 from anomalib.data.errors import MisMatchError
 from anomalib.data.utils import (
     DownloadInfo,
